Move `CardItem`, `card()` / `card_impl()`, `card_header()`, and `card_footer()`
